Fix GDPR usage
The newer version does not support model being
a subpackage. I ended up moving ProfileUser to
the models.py in api application.

Also, some packages are still using pytz so to
keep Django backward compatabile, I had to add
add USE_DEPRECATED_PYTZ setting.

Refs TILA-2582
